requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
onnxruntime==1.16.3
ultralytics==8.0.206
torch==2.1.1
torchvision==0.16.1
onnx==1.15.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 